refactor: split storage/database.py into domain modules
Function mapping:
clients.py: add_client, delete_client, client_has_reviews, list_clients, get_client
costs.py: log_cost, check_daily_limit, check_session_limit, get_session_cost
playbooks.py: 12 playbook/version functions
graphs.py: save_graph, load_graph, _ensure_graph_tables
recovery.py: save_recovery_state, load_recovery_state, delete_recovery_state
reviews.py: save_review_report, load_review_report, list_recent_reviews, list_reviews_for_client
comparisons.py: record_comparison, list_comparison_history, _ensure_comparison_history_table
search.py: search_all
database.py: keeps get_connection, transaction, init_database, run_migrations, _exec_migration_safely
