+++ src/openreview_cli/tui/domain/search.py
@@ -10,7 +10,7 @@ def search_all_via_tui(query: str) -> dict[str, list[dict[str, Any]]]:
 
     Results are grouped by type: ``reviews``, ``clients``, ``playbooks``.
     """
-    from openreview_cli.storage.database import search_all as _search_all
+    from openreview_cli.storage.search import search_all as _search_all
     from openreview_cli.tui.domain.clients import get_db_path
 
     return _search_all(get_db_path(), query)
