+++ tests/unit/test_playbook_history.py
@@ -7,10 +7,12 @@ from pathlib import Path
 import pytest
 
 from openreview_cli.storage.database import (
+    init_database,
+)
+from openreview_cli.storage.playbooks import (
     delete_playbook,
     get_playbook_history,
     import_playbook_yaml,
-    init_database,
     set_current_version,
 )
